requirements.txt
opencv-python
ivy-core
ivy-mech